move: CoordinateComponent method move
