+++ api/src/main/kotlin/io/github/sakaki_aruka/customcrafter/util/Converter.kt
@@ -1,61 +1,12 @@
 package io.github.sakaki_aruka.customcrafter.util
 
-import io.github.sakaki_aruka.customcrafter.recipe.CoordinateComponent
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.minimessage.MiniMessage
 
+/**
+ * @suppress
+ */
 object Converter {
-    /**
-     * ※ This is a high cost function.
-     *
-     * ```kotlin
-     * val list = CoordinateComponent.squareFill(3)
-     * println(Converter.getComponentsShapeString(list))
-     * // -> OUTPUT
-     * // ###
-     * // ###
-     * // ###
-     * ```
-     *
-     * ```kotlin
-     * val list = CoordinateComponent.square(3)
-     * println(Converter.getComponentsShapeString(list))
-     * // -> OUTPUT
-     * // ###
-     * // #_#
-     * // ###
-     * ```
-     *
-     * @param[list] List of coordinates
-     * @param[existsSlotChar] This is a char what is used on coordinate exists. Default = `#`.
-     * @param[notExistsSlotChar] This is a char what is used on coordinate NOT exists. Default = `_`.
-     * @return[String] Shape
-     * @since 5.0.16
-     */
-    @JvmStatic
-    @JvmOverloads
-    fun getComponentsShapeString(
-        list: Collection<CoordinateComponent>,
-        existsSlotChar: Char = '#',
-        notExistsSlotChar: Char = '_'
-    ): String {
-        val minX = list.minOf { it.x }
-        val minY = list.minOf { it.y }
-        val maxX = list.maxOf { it.x }
-        val maxY = list.maxOf { it.y }
-
-        val builder = StringBuilder()
-        (minY..maxY).forEach { y ->
-            (minX..maxX).forEach { x ->
-                builder.append(
-                    if (list.contains(CoordinateComponent(x, y))) existsSlotChar
-                    else notExistsSlotChar
-                )
-            }
-            if (y != maxY) builder.append(System.lineSeparator())
-        }
-        return builder.toString()
-    }
 
     /**
      * Kotlin string extension. Deserializes a MiniMessage string to a [Component] (net.kyori.adventure).
@@ -67,6 +18,6 @@ object Converter {
      * @return[Component] Deserialized component
      * @since 5.0.0
      */
-    fun String.toComponent(): Component = MiniMessage.miniMessage().deserialize(this)
+    internal fun String.toComponent(): Component = MiniMessage.miniMessage().deserialize(this)
 
 }
